+++ lib/src/clojure_lsp/feature/completion.clj
@@ -159,8 +159,12 @@
                                            documentation))))))
 
 (defn ^:private completion-item-with-alias-edit
-  [completion-item cursor-loc alias-to-add ns-to-add db]
-  (let [edits (some-> cursor-loc
+  [completion-item cursor-loc alias-to-add ns-to-add rcf-pos db]
+  (let [zloc (cond-> cursor-loc
+               rcf-pos edit/to-root
+               rcf-pos (edit/find-at-pos
+                         (:row rcf-pos) (:col rcf-pos)))
+        edits (some-> zloc
                       (f.add-missing-libspec/add-known-alias alias-to-add ns-to-add db)
                       r.transform/result)]
     (cond-> completion-item
@@ -178,15 +182,18 @@
 
 (defn ^:private completion-item-with-unresolved-alias-edit
   [completion-item cursor-loc alias-to-add ns-to-add db uri resolve-support]
-  (if (contains? (:properties resolve-support) "additionalTextEdits")
+  (let [rcf-pos (when (= :always (settings/get db [:add-missing :add-to-rcf]))
+                  (some-> cursor-loc edit/inside-rcf? z/node meta))]
+    (if (contains? (:properties resolve-support) "additionalTextEdits")
     ;; client supports postponing the expensive edit calculation
-    (add-unresolved completion-item
-                    "alias"
-                    {:ns-to-add    (name ns-to-add)
-                     :alias-to-add (name alias-to-add)
-                     :uri          uri})
+      (add-unresolved completion-item
+                      "alias"
+                      {:ns-to-add    (name ns-to-add)
+                       :alias-to-add (name alias-to-add)
+                       :rcf-pos rcf-pos
+                       :uri          uri})
     ;; client can't postpone the edit calculation, so do it now, even though it's expensive
-    (completion-item-with-alias-edit completion-item cursor-loc alias-to-add ns-to-add db)))
+      (completion-item-with-alias-edit completion-item cursor-loc alias-to-add ns-to-add rcf-pos db))))
 
 (defn ^:private generic-priority->specific-priority
   [element priority]
@@ -725,9 +732,9 @@
                                                                        :content-format-capability-path [:text-document :completion :completion-item :documentation-format]})
     item))
 
-(defmethod resolve-unresolved "alias" [_ item _other-unresolved {:keys [uri alias-to-add ns-to-add db]}]
+(defmethod resolve-unresolved "alias" [_ item _other-unresolved {:keys [uri alias-to-add ns-to-add rcf-pos db]}]
   (if-let [zloc (parser/safe-zloc-of-file db uri)]
-    (completion-item-with-alias-edit item zloc (symbol alias-to-add) (symbol ns-to-add) db)
+    (completion-item-with-alias-edit item zloc (symbol alias-to-add) (symbol ns-to-add) rcf-pos db)
     item))
 
 (defn resolve-item [{:keys [data] :as item} db*]