+++ lib/src/clojure_lsp/feature/refactor.clj
@@ -24,17 +24,17 @@
 
 (defmulti refactor :refactoring)
 
-(defmethod ^{:deprecated "Use add-missing-import instead"} refactor :add-import-to-namespace [{:keys [loc uri args db]}]
-  (apply f.add-missing-libspec/add-missing-import loc uri (concat args [db])))
+(defmethod ^{:deprecated "Use add-missing-import instead"} refactor :add-import-to-namespace [{:keys [loc uri args db components]}]
+  (apply f.add-missing-libspec/add-missing-import loc uri (concat args [db components])))
 
-(defmethod refactor :add-missing-import [{:keys [loc uri db args]}]
-  (apply f.add-missing-libspec/add-missing-import loc uri (concat args [db])))
+(defmethod refactor :add-missing-import [{:keys [loc uri db args components]}]
+  (apply f.add-missing-libspec/add-missing-import loc uri (concat args [db components])))
 
-(defmethod refactor :add-missing-libspec [{:keys [loc uri db]}]
-  (f.add-missing-libspec/add-missing-libspec loc uri db))
+(defmethod refactor :add-missing-libspec [{:keys [loc uri db components]}]
+  (f.add-missing-libspec/add-missing-libspec loc uri db components))
 
-(defmethod refactor :add-require-suggestion [{:keys [loc uri args db]}]
-  (apply f.add-missing-libspec/add-require-suggestion loc uri (concat args [db])))
+(defmethod refactor :add-require-suggestion [{:keys [loc uri args db components]}]
+  (apply f.add-missing-libspec/add-require-suggestion loc uri (concat args [db components])))
 
 (defmethod refactor :clean-ns [{:keys [loc uri db]}]
   (f.clean-ns/clean-ns-edits loc uri db))